% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics_extender.R
\docType{data}
\name{basics_extender}
\alias{basics_extender}
\title{Basics extender functions list}
\format{
An object of class \code{list} of length 2.
}
\usage{
basics_extender
}
\description{
List containing functions to extend \code{json_section}
}
\keyword{datasets}
