% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_json_list.R
\name{build_json_list}
\alias{build_json_list}
\title{Build JSON list from sections}
\usage{
build_json_list(form_data, sections)
}
\arguments{
\item{form_data}{list of inputs from front end}

\item{sections}{list of section R6 objects}
}
\value{
list
}
\description{
Build a list ready for JSON parsing
}
