% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_section_object.R
\name{create_section_object}
\alias{create_section_object}
\title{Create JSON section R6 object}
\usage{
create_section_object(name, extender)
}
\arguments{
\item{name}{R6 class name, will be used as the section name}

\item{extender}{list of functions including input_group_ui and value_extractor}
}
\value{
an R6 object
}
\description{
Extend the \code{json_section} R6 to include specifics for a section
}
