% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_input.R
\name{date_input}
\alias{date_input}
\title{Date input}
\usage{
date_input(id, label, value = NULL)
}
\arguments{
\item{id}{The \code{input} slot that will be used to access the value.}

\item{label}{Display label for the control, or NULL for no label.}

\item{value}{The starting date. Either a Date object, or a string in yyyy-mm-dd format.
If NULL (the default), will use the current date in the client's time zone}
}
\value{
html tag
}
\description{
Date input
}
