% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_list_parser.R
\name{form_list_parser}
\alias{form_list_parser}
\title{Parse data captured from the JSON inputs}
\usage{
form_list_parser(form_list)
}
\arguments{
\item{form_list}{a named list}
}
\value{
a list
}
\description{
A named list is captured from the front end. This function parses the list given
the names e.g. basics_name, basics_profiles_0_username to correctly structure an
output list to parse into JSON. Where elements need to be unnamed in the final output,
they will be numbered here e.g. basics$profiles$\code{0}$username, and thus will need to be
unnamed downstream e.g \code{names(this$basics$profiles) <- NULL}
}
