% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3_collapse.R
\name{h3_collapse}
\alias{h3_collapse}
\title{Create h3 with collapsable inner}
\usage{
h3_collapse(title, ...)
}
\arguments{
\item{title}{title text}

\item{...}{inner elements}
}
\description{
A h3 title with inners collapsable. Uses the JS script ...
}
