% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_group_div.R
\name{input_group_div}
\alias{input_group_div}
\title{Container for group inputs}
\usage{
input_group_div(group_name, id_suffix, ...)
}
\arguments{
\item{group_name}{section name to identify the group}

\item{id_suffix}{unique value to append to id to make uniquely identifiable}

\item{...}{elements to include in div}
}
\description{
Contain for grouped inputs
}
