% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_downloadHandler.R
\name{json_downloadHandler}
\alias{json_downloadHandler}
\title{Download handler for exporting JSON}
\usage{
json_downloadHandler(json_list)
}
\arguments{
\item{json_list}{list that will be parsed into JSON}
}
\description{
Download handler for exporting JSON
}
