% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_list_parser.R
\name{json_list_parser}
\alias{json_list_parser}
\title{Parse a JSON resume list}
\usage{
json_list_parser(list)
}
\arguments{
\item{list}{JSON resume parsed list}
}
\description{
Adds numbers to unnamed repeatable elements that when unlisted will match
input ids on the front end. Assume repeatable fields have been entered into
options() (the UI components do this) ..
}
