% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_section.R
\name{json_section}
\alias{json_section}
\title{Create an JSON section R6 object}
\description{
Create an object that has a number of generic methods for creating JSON resume
input sections, retrieving input values and updating values.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{section_name}}{character, section name from \code{initialize}}

\item{\code{section_id}}{character, lower case section name}

\item{\code{id_counter}}{numeric, the count of inputs created for a section.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-json_section-new}{\code{json_section$new()}}
\item \href{#method-json_section-add_section_container}{\code{json_section$add_section_container()}}
\item \href{#method-json_section-input_group_ui}{\code{json_section$input_group_ui()}}
\item \href{#method-json_section-insert_section_ui}{\code{json_section$insert_section_ui()}}
\item \href{#method-json_section-clear_inputs}{\code{json_section$clear_inputs()}}
\item \href{#method-json_section-import_values}{\code{json_section$import_values()}}
\item \href{#method-json_section-observe_add_section}{\code{json_section$observe_add_section()}}
\item \href{#method-json_section-get_values}{\code{json_section$get_values()}}
\item \href{#method-json_section-value_extractor}{\code{json_section$value_extractor()}}
\item \href{#method-json_section-clone}{\code{json_section$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-json_section-new"></a>}}
\if{latex}{\out{\hypertarget{method-json_section-new}{}}}
\subsection{Method \code{new()}}{
Initialize an \code{json_section} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{json_section$new(section_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{section_name}}{The name of the section. This will be used as the prefix for many
tag IDs, text, as well as to match to the JSON schema.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-json_section-add_section_container"></a>}}
\if{latex}{\out{\hypertarget{method-json_section-add_section_container}{}}}
\subsection{Method \code{add_section_container()}}{
Create a HTML container for section inputs. Expects that the method \code{input_group_ui}
exists
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{json_section$add_section_container()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-json_section-input_group_ui"></a>}}
\if{latex}{\out{\hypertarget{method-json_section-input_group_ui}{}}}
\subsection{Method \code{input_group_ui()}}{
A slot for the input group UI. Expected that this class will be extended
and this method updated then.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{json_section$input_group_ui(id_suffix, values = NULL, hide = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id_suffix}}{a number to append to tag IDs - use \code{self$id_counter}}

\item{\code{values}}{a list of values to fill inputs}

\item{\code{hide}}{whether to create the UI hidden}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-json_section-insert_section_ui"></a>}}
\if{latex}{\out{\hypertarget{method-json_section-insert_section_ui}{}}}
\subsection{Method \code{insert_section_ui()}}{
Insert self$input_group_ui into section container
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{json_section$insert_section_ui(values = NULL, hide = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{values}}{a list of values to fill inputs}

\item{\code{hide}}{whether to include display:none as an inline style}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-json_section-clear_inputs"></a>}}
\if{latex}{\out{\hypertarget{method-json_section-clear_inputs}{}}}
\subsection{Method \code{clear_inputs()}}{
Remove all inputs for a current section.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{json_section$clear_inputs()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-json_section-import_values"></a>}}
\if{latex}{\out{\hypertarget{method-json_section-import_values}{}}}
\subsection{Method \code{import_values()}}{
Update input values from that imported from a JSON resume. Assumes \code{jsonlite::fromJSON()}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{json_section$import_values(values)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{values}}{values a list of named values}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-json_section-observe_add_section"></a>}}
\if{latex}{\out{\hypertarget{method-json_section-observe_add_section}{}}}
\subsection{Method \code{observe_add_section()}}{
Add a shiny observer to links the 'add *' to self$insert_section_ui()
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{json_section$observe_add_section(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{shiny input list}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-json_section-get_values"></a>}}
\if{latex}{\out{\hypertarget{method-json_section-get_values}{}}}
\subsection{Method \code{get_values()}}{
Retrieve values from the frontend input sectins
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{json_section$get_values(form_data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{form_data}}{list from JS that returns input values}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-json_section-value_extractor"></a>}}
\if{latex}{\out{\hypertarget{method-json_section-value_extractor}{}}}
\subsection{Method \code{value_extractor()}}{
A slot for a function that extracts name values and returns a named list for
a specific JSON resume section. It is expected that this class will be extended
and this method updated then.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{json_section$value_extractor(values)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{values}}{list from JS that returns input values}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-json_section-clone"></a>}}
\if{latex}{\out{\hypertarget{method-json_section-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{json_section$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
