% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonResumeEditR-package.R
\docType{package}
\name{jsonResumeEditR-package}
\alias{jsonResumeEditR}
\alias{jsonResumeEditR-package}
\title{jsonResumeEditR: R Shiny App To Edit A JSON Resume}
\description{
An R Shiny app to create, edit and maintain an JSON Resume.
}
\author{
\strong{Maintainer}: Shaun Nielsen \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
