% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_repeatable.R
\name{make_repeatable}
\alias{make_repeatable}
\title{Create a repeatable form section}
\usage{
make_repeatable(link_to, ...)
}
\arguments{
\item{link_to}{section id to link to e.g. basics_profiles, work}

\item{...}{form inputs}
}
\description{
Create a repeatable form section
}
