% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_template.R
\name{mod_template}
\alias{mod_template}
\title{Template server for JSON resume information}
\usage{
mod_template(id, do_what = c("clear", "import", "retrieve"), import)
}
\arguments{
\item{id}{id to use for shiny namespacing}

\item{do_what}{character, flag to indicate what to do with inputs: clear, import (used with \code{import}) or retrieve}

\item{import}{list, imported JSON resume used for populating inputs}
}
\description{
Template server for JSON resume information
}
