% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_empty_inputs.R
\name{remove_empty_inputs}
\alias{remove_empty_inputs}
\title{Remove empty inputs from captured values}
\usage{
remove_empty_inputs(x)
}
\arguments{
\item{x}{list of values}
}
\description{
Remove input values of '' from captured list of values
}
