% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_group_button.R
\name{remove_group_button}
\alias{remove_group_button}
\title{Action button to remove input fields}
\usage{
remove_group_button()
}
\value{
HTML tag
}
\description{
Runs custom javascript to remove input fields
}
