% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabPanel_basics.R
\name{tabPanel_basics}
\alias{tabPanel_basics}
\title{TabPanel for JSON resume basics information}
\usage{
tabPanel_basics(id)
}
\arguments{
\item{id}{id to use for shiny namespacing}
}
\description{
TabPanel for JSON resume basics information
}
