% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabPanel_template.R
\name{tabPanel_template}
\alias{tabPanel_template}
\title{TabPanel template for JSON resume basics information}
\usage{
tabPanel_template(id)
}
\arguments{
\item{id}{id to use for shiny namespacing}
}
\description{
TabPanel template for JSON resume basics information
}
