% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_input.R
\name{text_input}
\alias{text_input}
\title{Text input}
\usage{
text_input(id, label, value = NULL, placeholder = NULL, width = NULL)
}
\arguments{
\item{id}{The input slot that will be used to access the value.}

\item{label}{Display label for the control, or NULL for no label.}

\item{value}{Initial value.}

\item{placeholder}{A character string giving the user a hint as to what can be
entered into the control. Internet Explorer 8 and 9 do not support this option.}

\item{width}{The width of the input, e.g. '400px', or '100\%'}
}
\description{
Create a text input
}
