% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textarea_input.R
\name{textarea_input}
\alias{textarea_input}
\title{Text area input}
\usage{
textarea_input(id, label, value = NULL, placeholder = NULL)
}
\arguments{
\item{id}{The input slot that will be used to access the value.}

\item{label}{Display label for the control, or NULL for no label.}

\item{value}{Initial value.'}

\item{placeholder}{A character string giving the user a hint as to what can be
entered into the control. Internet Explorer 8 and 9 do not support this option.}
}
\description{
Create a text input
}
