% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cards_ui.R
\name{theme_cards_ui}
\alias{theme_cards_ui}
\title{Create cards show markup template}
\usage{
theme_cards_ui()
}
\description{
Create card to show previews of templates inlcuded in the package. There is
an onclick function that the Shiny server will use to wrangle the template and
display in an iframe
}
