% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_preview_iframe.R
\name{theme_preview_iframe}
\alias{theme_preview_iframe}
\title{Create an iframe for previewing (built) templates}
\usage{
theme_preview_iframe()
}
\value{
tagList
}
\description{
Create an iframe for previewing (built) templates
}
