% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unlist_and_rename}
\alias{unlist_and_rename}
\title{Unlist a list and convert . to _}
\usage{
unlist_and_rename(list)
}
\arguments{
\item{list}{a list}
}
\description{
List elements and HTML ID can be easily matched if we prepare well. Unlist
a list generates '.' in nested elements, and we want them to be '_' to match
HTML IDs were prepared eariler.
}
