% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_input_values.R
\name{update_input_values}
\alias{update_input_values}
\title{Update input values}
\usage{
update_input_values(x)
}
\arguments{
\item{x}{named vector with names matching HMTL IDs}
}
\description{
Update input values using a named vector that matches HTML IDs using jQuery
}
