% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_json_list.R
\name{build_json_list}
\alias{build_json_list}
\title{Build JSON list from sections}
\usage{
build_json_list(form_data, sections)
}
\arguments{
\item{form_data}{list of inputs from front end}

\item{sections}{list of section R6 objects}
}
\value{
list
}
\description{
getFormData() returns a named list of values e.g. \verb{$basics-name-1}, \verb{$profiles-network-2}.
The prefix is used to segregate sections (e.g. basics vs work), while the suffix used to group
related sub-sections (e.g. multiple work inputs). If there are empty sections, then their
return values will be "".

Each section object has methods to select the relevant values and wrangle them
}
\details{
Build a list ready for JSON parsing
}
