% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_jsonResumeEditR.R
\name{run_jsonResumeEditR}
\alias{run_jsonResumeEditR}
\title{Run jsonResumeEditR}
\usage{
run_jsonResumeEditR(testDir = FALSE)
}
\arguments{
\item{testDir}{logical, test whether function can find the app directory and skip starting the app. Used in testing.}
}
\description{
Run jsonResumeEditR shipped with the R package
}
